fix: close 3 more numpy.float64 leaks found via real ISI reference footage
Running real official ISI Freestyle 1/2/3 test videos through the pipeline
(uploaded by the user) surfaced np.float64(...) still showing up in the
jump/spin repr despite the prior audit — the audit's own manual review
missed these three, which real data caught immediately:

  - j['airtime'] via `times[i] - times[jump_start_idx]` (times is a
    numpy array)
  - sp['duration'] via the same pattern in _confirm_and_add_spin
  - out['axis_stability'] via `1.0 - np.std(xs) * 20`

Same fix as before: float(...) at the point each value leaves numpy.
Verified end-to-end: re-ran a real reference video, jumps/spins no longer
show np.float64(...) in repr, and strict json.dumps() (no default=
fallback) succeeds without the TypeError this bug class caused in
production. This is exactly why the project's own rule is to validate
detector changes against real footage rather than trust code review alone
— the previous audit looked correct and still missed three instances of
the same bug in the same file.

+++ src/analysis/ai_movement_engine.py
@@ -145,7 +145,9 @@ class AIMovementEngine:
 
         # Axis stability
         xs = [_hip_x(p) for p in seg if _hip_x(p) is not None]
-        out['axis_stability'] = round(max(0.0, 1.0 - np.std(xs) * 20), 2) if xs else 0.5
+        # float(...): np.std() returns numpy.float64, which stays numpy-typed
+        # through round()/max() and leaks into json.dumps() downstream.
+        out['axis_stability'] = round(float(max(0.0, 1.0 - np.std(xs) * 20)), 2) if xs else 0.5
 
         # MovementClassifier
         if self._mc and self._MF:
